python/cloud_image_api/requirements.txt
fastapi
google-cloud-vision
uvicorn
